Break up test_moments.py
