+++ pymc/tests/distributions/test_mixture.py
@@ -62,9 +62,8 @@ from pymc.sampling import (
     sample_prior_predictive,
 )
 from pymc.step_methods import Metropolis
-from pymc.tests.distributions.test_moments import assert_moment_is_expected
 from pymc.tests.distributions.test_random import pymc_random
-from pymc.tests.distributions.util import Domain, Simplex
+from pymc.tests.distributions.util import Domain, Simplex, assert_moment_is_expected
 from pymc.tests.helpers import SeededTest
